SFORM-50 entityからdao相当部分を分離
