+++ adminApi/app/net/macolabo/sform2/modules/BaseModule.scala
@@ -2,7 +2,7 @@ package net.macolabo.sform2.modules
 
 import com.google.inject.AbstractModule
 import net.codingwell.scalaguice.ScalaModule
-import net.macolabo.sform2.domain.models.daos.{ApiTokenDAO, ApiTokenDAOImpl, AuthTokenDAO, AuthTokenDAOImpl, FormDAO, FormDAOImpl, PostdataDAO, PostdataDAOImpl, TransferDetailLogDAO, TransferDetailLogDAOImpl, TransferLogDAO, TransferLogDAOImpl, UserDAO, UserDAOImpl}
+import net.macolabo.sform2.domain.models.daos.{ApiTokenDAO, ApiTokenDAOImpl, AuthTokenDAO, AuthTokenDAOImpl, FormDAO, FormDAOImpl, PostdataDAO, PostdataDAOImpl, TransferConfigDAO, TransferConfigDAOImpl, TransferConfigMailAddressDAO, TransferConfigMailAddressDAOImpl, TransferConfigSalesforceDAO, TransferConfigSalesforceDAOImpl, TransferConfigSalesforceObjectDAO, TransferConfigSalesforceObjectDAOImpl, TransferConfigSalesforceObjectFieldDAO, TransferConfigSalesforceObjectFieldDAOImpl, TransferDetailLogDAO, TransferDetailLogDAOImpl, TransferLogDAO, TransferLogDAOImpl, UserDAO, UserDAOImpl}
 import net.macolabo.sform2.domain.services.ApiToken.{ApiTokenService, ApiTokenServiceImpl}
 import net.macolabo.sform2.domain.services.AuthToken.{AuthTokenService, AuthTokenServiceImpl}
 import net.macolabo.sform2.domain.services.GoogleAuth.{GoogleAuthService, GoogleAuthServiceImpl}
@@ -28,5 +28,10 @@ class BaseModule extends AbstractModule with ScalaModule {
     bind[ApiTokenService].to[ApiTokenServiceImpl]
     bind[UserDAO].to[UserDAOImpl]
     bind[GoogleAuthService].to[GoogleAuthServiceImpl]
+    bind[TransferConfigMailAddressDAO].to[TransferConfigMailAddressDAOImpl]
+    bind[TransferConfigSalesforceDAO].to[TransferConfigSalesforceDAOImpl]
+    bind[TransferConfigSalesforceObjectDAO].to[TransferConfigSalesforceObjectDAOImpl]
+    bind[TransferConfigSalesforceObjectFieldDAO].to[TransferConfigSalesforceObjectFieldDAOImpl]
+    bind[TransferConfigDAO].to[TransferConfigDAOImpl]
   }
 }
